refactor : 가독성 향상을 위한 메서드 순서 정렬 및 메서드, 변수명 개선
