--- src/main/java/domain/Ladder.java
@@ -1,34 +1,0 @@
-package domain;
-
-import java.util.Collections;
-import java.util.List;
-import java.util.Objects;
-
-public class Ladder {
-    private final List<Line> lines;
-
-    public Ladder(final List<Line> lines) {
-        this.lines = lines;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Ladder ladder = (Ladder) o;
-        return Objects.equals(lines, ladder.lines);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(lines);
-    }
-
-    public List<Line> getLines() {
-        return Collections.unmodifiableList(lines);
-    }
-}
